// src/types/employee.ts
export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  departmentId: number;
  position: string;
  jobId?: number; // Reference to the job opening
  employeeId?: string; // Optional as it's generated by backend
  status: 'ACTIVE' | 'INACTIVE' | 'ON_LEAVE';
  joinDate: string;
  createdAt: string;
  phoneNumber: string;
  address: string;
  emergencyContact: {
    name: string;
    relationship: string;
    phoneNumber: string;
  };
  bankDetails: {
    accountName: string;
    accountNumber: string;
    bankName: string;
    branchName: string;
  };
  documents: {
    type: string;
    number: string;
    expiryDate?: string;
  }[];
} 